Restructure codegen pipeline loading & running code, in preparation for even more refactoring
