+++ cmd/cli/codegen/kindsyscomposable.go
@@ -1,4 +1,4 @@
-package loaders
+package codegen
 
 import (
 	"fmt"
