+++ cmd/cli/codegen/kindsyscore.go
@@ -1,4 +1,4 @@
-package loaders
+package codegen
 
 import (
 	"path/filepath"
