--- cmd/cli/loaders/tools.go
@@ -1,54 +1,0 @@
-package loaders
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"path/filepath"
-	"strings"
-)
-
-func guessPackageFromFilename(filename string) string {
-	pkg := filepath.Base(filepath.Dir(filename))
-	if pkg != "." {
-		return pkg
-	}
-
-	return strings.TrimSuffix(filepath.Base(filename), filepath.Ext(filename))
-}
-
-func dirExists(dir string) (bool, error) {
-	stat, err := os.Stat(dir)
-	//nolint:gocritic
-	if os.IsNotExist(err) {
-		return false, nil
-	} else if err != nil {
-		return false, err
-	} else if !stat.IsDir() {
-		return false, fmt.Errorf("'%s' is not a directory", dir)
-	}
-
-	return true, nil
-}
-
-func loadURL(ctx context.Context, url string) (io.ReadCloser, error) {
-	client := http.DefaultClient
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("expecting 200 when loading '%s', got %d", url, resp.StatusCode)
-	}
-
-	return resp.Body, nil
-}
