+++ internal/jennies/python/jennies.go
@@ -28,6 +28,10 @@ func New(config Config) *Language {
 	}
 }
 
+func (language *Language) Name() string {
+	return LanguageRef
+}
+
 func (language *Language) Jennies(globalConfig common.Config) *codejen.JennyList[common.Context] {
 	jenny := codejen.JennyListWithNamer[common.Context](func(_ common.Context) string {
 		return LanguageRef
